chore(analytics-core): move core config interface under types
