+++ packages/analytics-core/src/index.ts
@@ -3,7 +3,8 @@ export { Identify, IIdentify } from './identify';
 export { Revenue, IRevenue, RevenueProperty } from './revenue';
 export { Destination } from './plugins/destination';
 export { IdentityEventSender } from './plugins/identity';
-export { IConfig, Config, RequestMetadata } from './config';
+export { Config, RequestMetadata } from './config';
+export { IConfig } from './types/config/core';
 export { Logger, ILogger, LogConfig } from './logger';
 export { getGlobalScope } from './global-scope';
 export { getAnalyticsConnector, setConnectorDeviceId, setConnectorUserId } from './analytics-connector';
